requirements.txt
flask>=3.0.0,<4
flask-mail>=0.9.1,<1
flask-login>=0.6.3,<1
gunicorn>=21.0,<23
google-generativeai>=0.8.0,<1
werkzeug>=3.0.0,<4
